I follow single responsibility principle to create new, modular, libs.
