+++ lib/utils/chunkArray.ts
@@ -1,0 +1,12 @@
+/**
+ * Splits an array into smaller chunks of specified size
+ */
+export const chunkArray = <T>(array: T[], size: number): T[][] => {
+  const chunks: T[][] = [];
+  for (let i = 0; i < array.length; i += size) {
+    chunks.push(array.slice(i, i + size));
+  }
+  return chunks;
+};
+
+export default chunkArray;
