+++ types/agent.ts
@@ -6,6 +6,7 @@ export type SocialPost = Database["public"]["Tables"]["social_posts"]["Row"];
 export type Agent = Database["public"]["Tables"]["agents"]["Row"];
 export type Account_Social =
   Database["public"]["Tables"]["account_socials"]["Row"];
+export type Comment = Database["public"]["Tables"]["post_comments"]["Row"];
 
 export type ScrapedComment = {
   comment: string | null;
